all screens added to component files
