+++ frontend/src/App.js
@@ -3,31 +3,31 @@ import {BrowserRouter as Router, Route, Switch,} from 'react-router-dom';
 import {Container} from 'react-bootstrap'
 import Header from "./components/HeaderFooter/Header";
 import Footer from "./components/HeaderFooter/Footer";
-import ProductsScreen from "./screens/Product(s)/ProductsScreen";
-import ProductScreen from "./screens/Product(s)/ProductScreen";
-import CartScreen from "./screens/Payment/CartScreen";
-import LoginScreen from "./screens/Login-Register/LoginScreen";
-import RegisterScreen from "./screens/Login-Register/RegisterScreen";
-import ProfileScreen from "./screens/ProfileScreen";
-import ShippingScreen from "./screens/Payment/ShippingScreen";
-import PaymentScreen from "./screens/Payment/PaymentScreen";
-import PlaceOrderScreen from "./screens/Payment/PlaceOrderScreen";
-import OrderScreen from "./screens/Payment/OrderScreen";
-import UserListScreen from "./screens/AdminScreens/UserListScreen";
-import UserEditScreen from "./screens/AdminScreens/UserEditScreen";
-import ProductListScreen from "./screens/AdminScreens/ProductListScreen";
-import ProductEditScreen from "./screens/AdminScreens/ProductEditScreen";
-import OrderListScreen from "./screens/AdminScreens/OrderListScreen";
-import ThankYouScreen from "./screens/Payment/ThankYouScreen";
+import ProductsScreen from "./components/Product(s)/ProductsScreen";
+import ProductScreen from "./components/Product(s)/ProductScreen";
+import CartScreen from "./components/Cart/Payment/CartScreen";
+import LoginScreen from "./components/SignIn/LoginScreen";
+import RegisterScreen from "./components/SignIn/RegisterScreen";
+import ProfileScreen from "./components/Profile/ProfileScreen";
+import ShippingScreen from "./components/Cart/Payment/ShippingScreen";
+import PaymentScreen from "./components/Cart/Payment/PaymentScreen";
+import PlaceOrderScreen from "./components/Cart/Payment/PlaceOrderScreen";
+import OrderScreen from "./components/Cart/Payment/OrderScreen";
+import UserListScreen from "./components/AdminScreens/UserListScreen";
+import UserEditScreen from "./components/AdminScreens/UserEditScreen";
+import ProductListScreen from "./components/AdminScreens/ProductListScreen";
+import ProductEditScreen from "./components/AdminScreens/ProductEditScreen";
+import OrderListScreen from "./components/Cart/OrderListScreen";
+import ThankYouScreen from "./components/Cart/Payment/ThankYouScreen";
 import HomeScreen from "./components/Home/HomeScreen";
 import AvionicsScreen from "./components/Avionics/AvionicsScreen";
 import MaintenanceScreen from "./components/Maintenance/MaintenanceScreen";
 import FlirScreen from "./components/Flir/FlirScreen";
-import TeamScreen from "./screens/TeamScreen";
+import TeamScreen from "./components/Team/TeamScreen";
 import ContactScreen from "./components/Contact/ContactScreen";
-import TermsScreen from "./screens/TermsScreens/TermsScreen";
-import TermsFlirScreen from "./screens/TermsScreens/TermsFlirScreen";
-import WishListScreen from "./screens/WishListScreen/WishListScreen"
+import TermsScreen from "./components/Terms/TermsScreen";
+import TermsFlirScreen from "./components/Terms/TermsFlirScreen";
+import WishListScreen from "./components/Wishlist/WishListScreen"
 import HeaderDesktop from "./components/HeaderFooter/HeaderDesktop";
 import HeaderMobile from "./components/HeaderFooter/HeaderMobile";
 import {listProducts} from "./actions/productActions";
