+++ frontend/src/components/Cart/Payment/OrderScreen.js
@@ -4,11 +4,11 @@ import { PayPalButton } from 'react-paypal-button-v2'
 import {Link} from "react-router-dom";
 import { Row, Col, ListGroup, Image, Card, Button} from "react-bootstrap";
 import {useDispatch, useSelector} from "react-redux";
-import Message from "../../components/Message";
-import Loader from "../../components/Loader";
-import {getOrderDetails, payOrder, deliverOrder } from "../../actions/orderActions";
-import { ORDER_PAY_RESET, ORDER_DELIVER_RESET } from "../../constants/orderConstants";
-import {CART_RESET, CART_SAVE_SHIPPING_ADDRESS} from "../../constants/cartConstants";
+import Message from "../../Message";
+import Loader from "../../Loader";
+import {getOrderDetails, payOrder, deliverOrder } from "../../../actions/orderActions";
+import { ORDER_PAY_RESET, ORDER_DELIVER_RESET } from "../../../constants/orderConstants";
+import {CART_RESET, CART_SAVE_SHIPPING_ADDRESS} from "../../../constants/cartConstants";
 
 const OrderScreen = ({ match, history }) => {
     const orderId = match.params.id;
