+++ frontend/src/components/Cart/Payment/PaymentScreen.js
@@ -1,9 +1,9 @@
 import React, {useState} from "react";
 import {Form, Button, Col} from 'react-bootstrap'
 import {useDispatch, useSelector} from "react-redux";
-import { savePaymentMethod } from "../../actions/cartActions";
-import FormContainer from "../../components/FormContainer";
-import CheckoutSteps from "../../components/Checkout/CheckoutSteps";
+import { savePaymentMethod } from "../../../actions/cartActions";
+import FormContainer from "../../FormContainer";
+import CheckoutSteps from "../CheckoutSteps";
 
 const PaymentScreen = ({ history }) => {
