+++ frontend/src/components/Cart/Payment/PlaceOrderScreen.js
@@ -2,9 +2,9 @@ import React, { useEffect } from "react";
 import {Link} from "react-router-dom";
 import {Button, Row, Col, ListGroup, Image, Card} from "react-bootstrap";
 import {useDispatch, useSelector} from "react-redux";
-import Message from "../../components/Message";
-import CheckoutSteps from "../../components/Checkout/CheckoutSteps";
-import {createOrder} from "../../actions/orderActions";
+import Message from "../../Message";
+import CheckoutSteps from "../CheckoutSteps";
+import {createOrder} from "../../../actions/orderActions";
 
 const PlaceOrderScreen = ({ history }) => {
     const cart = useSelector(state => state.cart);
