+++ frontend/src/components/Cart/Payment/ShippingScreen.js
@@ -1,9 +1,9 @@
 import React, {useState} from 'react'
 import {Form, Button} from 'react-bootstrap'
 import {useDispatch, useSelector} from 'react-redux'
-import FormContainer from "../../components/FormContainer";
-import CheckoutSteps from "../../components/Checkout/CheckoutSteps";
-import { saveShippingAddress } from "../../actions/cartActions";
+import FormContainer from "../../FormContainer";
+import CheckoutSteps from "../CheckoutSteps";
+import { saveShippingAddress } from "../../../actions/cartActions";
 
 const ShippingScreen = ({ history }) => {
