+++ frontend/src/components/GarminAd.js
@@ -3,7 +3,7 @@ import Fade from 'react-reveal/Fade';
 import {Col, Row} from "react-bootstrap";
 import AvionicsProducts from "./Avionics/AvionicsProducts";
 import '../style/2/GarminHomeFlag.scss'
-import ProductCarousel from "./Products/ProductsCarousel/ProductCarousel";
+import ProductCarousel from "./Product(s)/ProductsCarousel/ProductCarousel";
 
 
 const GarminAd = () => {
