+++ frontend/src/components/Profile/ProfileScreen.js
@@ -1,8 +1,8 @@
 import React, {useLayoutEffect} from 'react'
 import {Row, Col, Tab, Nav, Jumbotron, Container} from 'react-bootstrap'
-import ProfileEdit from "../components/Profile/ProfileEdit";
-import ProfileSavedWishList from "../components/Profile/ProfileSavedWishList";
-import '../style/ProfileScreen.css'
+import ProfileEdit from "./ProfileEdit";
+import ProfileSavedWishList from "./ProfileSavedWishList";
+import '../../style/ProfileScreen.css'
 import {useSelector} from "react-redux";
 
 const ProfileScreen = ({location, history}) => {
