+++ frontend/src/components/SignIn/RegisterScreen.js
@@ -2,9 +2,9 @@ import React, {useState, useEffect, useLayoutEffect} from 'react'
 import {Link} from 'react-router-dom'
 import {Form, Button, Row, Col, Container} from 'react-bootstrap'
 import {useDispatch, useSelector} from 'react-redux'
-import Message from '../../components/Message'
-import Loader from '../../components/Loader'
-import FormContainer from "../../components/FormContainer";
+import Message from '../Message'
+import Loader from '../Loader'
+import FormContainer from "../FormContainer";
 import {register} from '../../actions/userActions'
 import '../../style/registerScreen/registerScreen.css'
